requirements.txt
tpu-star==0.0.1-rc9
torch==1.6.0+cu101
torchvision==0.7.0+cu101
pre-commit==2.10.1
bezier==2020.5.19
augmixations==0.1.2
albumentations==0.5.2
neptune-client==0.5.1
pandas==1.2.2
opencv-python==4.4.0.46
psutil==5.8.0
tqdm==4.56.2